Add ConnectionInterface
We don't want to send data to external services when running tests i.e.
We don't want to open a TCP connection to a third-party or a service
when running our test suite. To solve this problem I have created a
simple ConnectionInterface object which is easily mockable when testing.